added routes, need to fix path bug having to do with router.use
