+++ 01_server/app/routes/sightings.router.js
@@ -1,0 +1,25 @@
+var router = require('express').Router();
+var mongoose = require('mongoose');
+
+var Sighting = mongoose.model('Sighting');
+
+router.get('/', function (req, res, next) {
+	Sighting.find()
+		.then(function (sightings) {
+			res.send(sightings);
+		}, next);
+});
+
+router.post('/', function (req, res, next) {
+	Sighting.create(req.body)
+		.then(function (result) {
+			res.send(result);
+		}, next);
+})
+
+router.get('/:id', function (req, res, next) {
+	Sighting.findById(req.params.id)
+		.then(function (sighting) {
+			res.send(sighting);
+		}, next)
+})
